+++ src/main/java/net/jodah/failsafe/Fallback.java
@@ -16,18 +16,18 @@
 package net.jodah.failsafe;
 
 import net.jodah.failsafe.event.ExecutionAttemptedEvent;
-import net.jodah.failsafe.function.CheckedConsumer;
-import net.jodah.failsafe.function.CheckedFunction;
-import net.jodah.failsafe.function.CheckedRunnable;
-import net.jodah.failsafe.function.CheckedSupplier;
+import net.jodah.failsafe.function.*;
 import net.jodah.failsafe.internal.util.Assert;
 
-import static net.jodah.failsafe.Functions.toFn;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+
+import static net.jodah.failsafe.Functions.*;
 
 /**
  * A Policy that handles failures using a fallback function or result.
  * <p>
- * Note: Fallback extends {@link FailurePolicy} which offers additional configuration.
+ * Note: Fallback extends {@link FailurePolicy} {@link PolicyListeners} which offer additional configuration.
  * </p>
  *
  * @param <R> result type
@@ -40,10 +40,11 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
   public static final Fallback<Void> VOID = new Fallback<>();
 
   private final CheckedFunction<ExecutionAttemptedEvent, R> fallback;
+  private final CheckedFunction<ExecutionAttemptedEvent, CompletableFuture<R>> fallbackStage;
   private boolean async;
 
-  private Fallback(){
-    fallback = null;
+  private Fallback() {
+    this(null, null, false);
   }
 
   /**
@@ -51,8 +52,10 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    *
    * @throws NullPointerException if {@code fallback} is null
    */
-  private Fallback(CheckedFunction<ExecutionAttemptedEvent, R> fallback, boolean async) {
-    this.fallback = Assert.notNull(fallback, "fallback");
+  private Fallback(CheckedFunction<ExecutionAttemptedEvent, R> fallback,
+    CheckedFunction<ExecutionAttemptedEvent, CompletableFuture<R>> fallbackStage, boolean async) {
+    this.fallback = fallback;
+    this.fallbackStage = fallbackStage;
     this.async = async;
   }
 
@@ -62,7 +65,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    * @throws NullPointerException if {@code fallback} is null
    */
   public static <R> Fallback<R> of(CheckedRunnable fallback) {
-    return new Fallback<>(toFn(Assert.notNull(fallback, "fallback")), false);
+    return new Fallback<>(toFn(Assert.notNull(fallback, "fallback")), null, false);
   }
 
   /**
@@ -72,7 +75,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   @SuppressWarnings("unchecked")
   public static <R> Fallback<R> of(CheckedSupplier<? extends R> fallback) {
-    return new Fallback<>(toFn((CheckedSupplier<R>) Assert.notNull(fallback, "fallback")), false);
+    return new Fallback<>(toFn((CheckedSupplier<R>) Assert.notNull(fallback, "fallback")), null, false);
   }
 
   /**
@@ -83,7 +86,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   @SuppressWarnings("unchecked")
   public static <R> Fallback<R> of(CheckedConsumer<ExecutionAttemptedEvent<? extends R>> fallback) {
-    return new Fallback<>(toFn(Assert.notNull((CheckedConsumer) fallback, "fallback")), false);
+    return new Fallback<>(toFn(Assert.notNull((CheckedConsumer) fallback, "fallback")), null, false);
   }
 
   /**
@@ -94,7 +97,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   @SuppressWarnings("unchecked")
   public static <R> Fallback<R> of(CheckedFunction<ExecutionAttemptedEvent<? extends R>, ? extends R> fallback) {
-    return new Fallback<>(Assert.notNull((CheckedFunction) fallback, "fallback"), false);
+    return new Fallback<>(Assert.notNull((CheckedFunction) fallback, "fallback"), null, false);
   }
 
   /**
@@ -109,7 +112,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
     Assert.notNull((CheckedFunction) fallback, "fallback");
     return new Fallback<>(e -> {
       throw fallback.apply(e);
-    }, false);
+    }, null, false);
   }
 
   /**
@@ -117,7 +120,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   @SuppressWarnings("rawtypes")
   public static <R> Fallback<R> of(R fallback) {
-    return new Fallback<>(toFn(fallback), false);
+    return new Fallback<>(toFn(fallback), null, false);
   }
 
   /**
@@ -126,7 +129,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    * @throws NullPointerException if {@code fallback} is null
    */
   public static <R> Fallback<R> ofAsync(CheckedRunnable fallback) {
-    return new Fallback<>(toFn(Assert.notNull(fallback, "fallback")), true);
+    return new Fallback<>(toFn(Assert.notNull(fallback, "fallback")), null, true);
   }
 
   /**
@@ -136,7 +139,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   @SuppressWarnings("unchecked")
   public static <R> Fallback<R> ofAsync(CheckedSupplier<? extends R> fallback) {
-    return new Fallback<>(toFn((CheckedSupplier<R>) Assert.notNull(fallback, "fallback")), true);
+    return new Fallback<>(toFn((CheckedSupplier<R>) Assert.notNull(fallback, "fallback")), null, true);
   }
 
   /**
@@ -147,7 +150,7 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   @SuppressWarnings("unchecked")
   public static <R> Fallback<R> ofAsync(CheckedConsumer<ExecutionAttemptedEvent<? extends R>> fallback) {
-    return new Fallback<>(toFn(Assert.notNull((CheckedConsumer) fallback, "fallback")), true);
+    return new Fallback<>(toFn(Assert.notNull((CheckedConsumer) fallback, "fallback")), null, true);
   }
 
   /**
@@ -158,11 +161,47 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   @SuppressWarnings("unchecked")
   public static <R> Fallback<R> ofAsync(CheckedFunction<ExecutionAttemptedEvent<? extends R>, ? extends R> fallback) {
-    return new Fallback<>(Assert.notNull((CheckedFunction) fallback, "fallback"), true);
+    return new Fallback<>(Assert.notNull((CheckedFunction) fallback, "fallback"), null, true);
   }
 
-  public R apply(R result, Throwable failure, ExecutionContext context) throws Exception {
-    return fallback.apply(new ExecutionAttemptedEvent<>(result, failure, context));
+  /**
+   * Returns the {@code fallback} to be executed if execution fails.
+   *
+   * @throws NullPointerException if {@code fallback} is null
+   */
+  public static <R> Fallback<R> ofStage(CheckedSupplier<? extends CompletionStage<R>> fallback) {
+    return new Fallback<>(null, (CheckedFunction) toFn(Assert.notNull(fallback, "fallback")), false);
+  }
+
+  /**
+   * Returns the {@code fallback} to be executed if execution fails. The {@code fallback} accepts an {@link
+   * ExecutionAttemptedEvent}.
+   *
+   * @throws NullPointerException if {@code fallback} is null
+   */
+  public static <R> Fallback<R> ofStage(
+    CheckedFunction<ExecutionAttemptedEvent<? extends R>, ? extends CompletionStage<R>> fallback) {
+    return new Fallback<>(null, Assert.notNull((CheckedFunction) fallback, "fallback"), false);
+  }
+
+  /**
+   * Returns the {@code fallback} to be executed asynchronously if execution fails.
+   *
+   * @throws NullPointerException if {@code fallback} is null
+   */
+  public static <R> Fallback<R> ofStageAsync(CheckedSupplier<? extends CompletionStage<R>> fallback) {
+    return new Fallback<>(null, (CheckedFunction) toFn(Assert.notNull(fallback, "fallback")), true);
+  }
+
+  /**
+   * Returns the {@code fallback} to be executed asynchronously if execution fails. The {@code fallback} accepts an
+   * {@link ExecutionAttemptedEvent}.
+   *
+   * @throws NullPointerException if {@code fallback} is null
+   */
+  public static <R> Fallback<R> ofStageAsync(
+    CheckedFunction<ExecutionAttemptedEvent<? extends R>, ? extends CompletionStage<R>> fallback) {
+    return new Fallback<>(null, Assert.notNull((CheckedFunction) fallback, "fallback"), true);
   }
 
   /**
@@ -170,6 +209,22 @@ public class Fallback<R> extends FailurePolicy<Fallback<R>, R> {
    */
   public boolean isAsync() {
     return async;
+  }
+
+  /**
+   * Returns the applied fallback result.
+   */
+  R apply(R result, Throwable failure, ExecutionContext context) throws Exception {
+    ExecutionAttemptedEvent<R> event = new ExecutionAttemptedEvent<>(result, failure, context);
+    return fallback != null ? fallback.apply(event) : fallbackStage.apply(event).get();
+  }
+
+  /**
+   * Returns a future applied fallback result.
+   */
+  CompletableFuture<R> applyStage(R result, Throwable failure, ExecutionContext context) throws Exception {
+    ExecutionAttemptedEvent<R> event = new ExecutionAttemptedEvent<>(result, failure, context);
+    return fallback != null ? CompletableFuture.completedFuture(fallback.apply(event)) : fallbackStage.apply(event);
   }
 
   @Override